+++ packages/advanced-logic/src/extensions/payment-network/address-based.ts
@@ -51,9 +51,7 @@ export default abstract class AddressBasedPaymentNetwork<
       throw Error('refundAddress is not a valid address');
     }
 
-    return super.createCreationAction(
-      creationParameters,
-    ) as ExtensionTypes.IAction<TCreationParameters>;
+    return super.createCreationAction(creationParameters);
   }
 
   /**
@@ -108,14 +106,6 @@ export default abstract class AddressBasedPaymentNetwork<
     };
   }
 
-  /**
-   * Applies a creation
-   *
-   * @param isValidAddress address validator function
-   * @param extensionAction action to apply
-   *
-   * @returns state of the extension created
-   */
   protected applyCreation(
     extensionAction: ExtensionTypes.IAction,
     timestamp: number,
@@ -254,12 +244,6 @@ export default abstract class AddressBasedPaymentNetwork<
     return copiedExtensionState;
   }
 
-  /**
-   * Validate the extension action regarding the currency and network
-   * It must throw in case of error
-   *
-   * @param request
-   */
   protected validate(
     request: RequestLogicTypes.IRequest,
     // eslint-disable-next-line @typescript-eslint/no-unused-vars