+++ packages/advanced-logic/src/extensions/payment-network/ethereum/input-data.ts
@@ -9,11 +9,6 @@ const supportedNetworks = ['mainnet', 'rinkeby'];
 
 /**
  * Implementation of the payment network to pay in ETH based on input data.
- * With this extension, one request can have two Ethereum addresses (one for payment and one for refund) and a specific value to give as input data
- * Every ETH ethereum transaction that reaches these addresses and has the correct input data will be interpreted as a payment or a refund.
- * The value to give as input data is the last 8 bytes of a salted hash of the requestId and the address: `last8Bytes(hash(requestId + salt + address))`:
- * The salt should have at least 8 bytes of randomness. A way to generate it is:
- *   `Math.floor(Math.random() * Math.pow(2, 4 * 8)).toString(16) + Math.floor(Math.random() * Math.pow(2, 4 * 8)).toString(16)`
  */
 export default class EthInputPaymentNetwork extends ReferenceBasedPaymentNetwork {
   public constructor(